requirements.txt
langchain>=0.2.17
langchain-community>=0.2.17
langchain-openai>=0.2.7
langchain-pinecone>=0.1.3
langchain-text-splitters>=0.2.2
pinecone>=7.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
tiktoken>=0.5.0
pandas
numpy